backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
jinja2==3.1.4
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
beautifulsoup4==4.12.3
scikit-learn==1.5.2
networkx==3.3
nltk==3.9.1
langdetect==1.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
